- The toolbar will now wrap if the horizontal size of the window is too small - Changed rgba_t::amp to take signed arguments
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@1191 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Structs.h
@@ -249,7 +249,7 @@ struct rgba_t {
 	}
 
 	// Amplify/fade colour components by absolute amounts
-	rgba_t amp(uint8_t R, uint8_t G, uint8_t B, uint8_t A) {
+	rgba_t amp(int R, int G, int B, int A) {
 		int nr = r + R;
 		int ng = g + G;
 		int nb = b + B;
